+++ packages/react/src/Form/lang.ts
@@ -1,8 +1,0 @@
-export const FormDefaultLang = {
-  submit: 'Submit',
-  required: 'This field is required',
-  string: 'This field must be a string',
-  number: 'This field must be a number',
-}
-
-export type FormLang = typeof FormDefaultLang